pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lexar-ai"
version = "1.1.0"
description = "LEXAR: Legal EXplainable Augmented Reasoner - A retrieval-augmented generation system for legal question answering with strict evidence grounding"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "LEXAR Contributors"}
]
keywords = [
    "legal",
    "nlp",
    "rag",
    "retrieval-augmented-generation",
    "question-answering",
    "explainable-ai",
    "evidence-grounding"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic"
]

dependencies = [
    "transformers>=4.30.0,<5.0.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.21.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
cpu = [
    "torch>=2.0.0",
]
gpu = [
    "torch>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]
server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]
all = [
    "lexar-ai[cpu,dev,server]",
]

[project.urls]
Homepage = "https://github.com/yourusername/legalrag"
Documentation = "https://github.com/yourusername/legalrag#readme"
Repository = "https://github.com/yourusername/legalrag"
Issues = "https://github.com/yourusername/legalrag/issues"

[project.scripts]
lexar = "lexar.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lexar*"]
exclude = ["backend*", "scripts*", "data*", "evaluation*"]

[tool.setuptools.package-data]
lexar = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.tox
  | build
  | dist
  | __pycache__
  | \.mypy_cache
  | \.pytest_cache
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --tb=short"
